rename std tests folder

+++ prusti-tests/tests/std/pass/default.rs
@@ -1,7 +1,7 @@
 use prusti_contracts::*;
 
 fn main() {
-    let default: (i32, i32) = Default::default();
+    let default: (i32, u64) = Default::default();
     assert!(default.0 == 0);
     assert!(default.1 == 0);
 }
